+++ src/dependency-builder/detectives/detective-stylable/index.js
@@ -1,25 +1,33 @@
 const stylable = require('stylable');
 module.exports = function(src, options = {}) {
   const css = stylable.safeParse(src);
-  const dependencies = [];
-  const importSpecifiers = {};
+  const dependencies = {};
+  const addDependency = (dependency) => {
+    if (!dependencies[dependency]) {
+      dependencies[dependency] = {};
+    }
+  };
+  const addImportSpecifier = (dependency, importSpecifier) => {
+    if (dependencies[dependency].importSpecifiers) {
+      dependencies[dependency].importSpecifiers.push(importSpecifier);
+    } else {
+      dependencies[dependency].importSpecifiers = [importSpecifier];
+    }
+  };
 
   css.walkRules((rule) => {
     const stFrom = rule.nodes.find(node => node.prop === '-st-from');
     if (!stFrom) return;
     const stFromValue = stFrom.value.replace(/["']/g, '');
-    dependencies.push(stFromValue);
+    addDependency(stFromValue);
     const stNamed = rule.nodes.find(node => node.prop === '-st-named');
     if (!stNamed) return;
     const specifierValue = {
       isDefault: false, // @todo,
       name: stNamed.value
     };
-    importSpecifiers[stFromValue]
-      ? importSpecifiers[stFromValue].push(specifierValue)
-      : importSpecifiers[stFromValue] = [specifierValue];
+    addImportSpecifier(stFromValue, specifierValue);
   });
-  options.importSpecifiers = importSpecifiers;
 
   return dependencies;
 };