+++ src/dependency-builder/detectives/detective-typescript/index.js
@@ -2,8 +2,8 @@
 * this file had been forked from https://github.com/pahen/detective-typescript
 */
 
-var Parser = require('typescript-eslint-parser');
-var Walker = require('node-source-walk');
+const Parser = require('typescript-eslint-parser');
+const Walker = require('node-source-walk');
 
 /**
  * Extracts the dependencies of the supplied TypeScript module
@@ -15,9 +15,21 @@ var Walker = require('node-source-walk');
 module.exports = function(src, options = {}) {
   options.parser = Parser;
 
-  var walker = new Walker(options);
+  const walker = new Walker(options);
 
-  var dependencies = [];
+  const dependencies = {};
+  const addDependency = (dependency) => {
+    if (!dependencies[dependency]) {
+      dependencies[dependency] = {};
+    }
+  };
+  const addImportSpecifier = (dependency, importSpecifier) => {
+    if (dependencies[dependency].importSpecifiers) {
+      dependencies[dependency].importSpecifiers.push(importSpecifier);
+    } else {
+      dependencies[dependency].importSpecifiers = [importSpecifier];
+    }
+  };
 
   if (typeof src === 'undefined') {
     throw new Error('src not given');
@@ -27,40 +39,37 @@ module.exports = function(src, options = {}) {
     return dependencies;
   }
 
-  var importSpecifiers = {};
   walker.walk(src, function(node) {
     switch (node.type) {
       case 'ImportDeclaration':
         if (node.source && node.source.value) {
-          dependencies.push(node.source.value);
+          const dependency = node.source.value;
+          addDependency(dependency);
 
           node.specifiers.forEach((specifier) => {
-            var specifierValue = {
+            const specifierValue = {
               isDefault: specifier.type === 'ImportDefaultSpecifier',
               name: specifier.local.name
             };
-            importSpecifiers[node.source.value]
-              ? importSpecifiers[node.source.value].push(specifierValue)
-              : importSpecifiers[node.source.value] = [specifierValue];
+            addImportSpecifier(dependency, specifierValue);
           });
         }
         break;
       case 'ExportNamedDeclaration':
       case 'ExportAllDeclaration':
         if (node.source && node.source.value) {
-          dependencies.push(node.source.value);
+          addDependency(node.source.value);
         }
         break;
       case 'TSExternalModuleReference':
         if (node.expression && node.expression.value) {
-          dependencies.push(node.expression.value);
+          addDependency(node.expression.value);
         }
         break;
       default:
-        return;
+        break;
     }
   });
-  options.importSpecifiers = importSpecifiers;
 
   return dependencies;
 };